pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "show-tracker"
version = "0.1.0"
description = "Automatic cross-platform media tracking: detects what you're watching and logs episode-level history."
requires-python = ">= 3.11"
license = "MIT"
readme = "README.md"
keywords = ["media", "tracking", "tv", "episodes", "activitywatch"]

dependencies = [
    "guessit>=3.8",
    "requests>=2.31",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "Pillow>=10.0",
    "python-Levenshtein>=0.25",
    "rapidfuzz>=3.6",
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "apscheduler>=3.10",
    "psutil>=5.9",
    "click>=8.1",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "easyocr>=1.7",
]
windows = [
    "winsdk>=1.0",
]
linux = [
    "dbus-next>=0.2.3",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
show-tracker = "show_tracker.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/show_tracker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["show_tracker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
